+++ core-ng/src/main/java/core/framework/impl/web/HTTPServer.java
@@ -75,6 +75,7 @@ public class HTTPServer {
         if (server != null) {
             logger.info("shutting down http server");
             shutdownHandler.shutdown();
+            handler.webSocketHandler.shutdown();
         }
     }