+++ core-ng/src/main/java/core/framework/web/websocket/Channel.java
@@ -1,0 +1,14 @@
+package core.framework.web.websocket;
+
+import java.util.Map;
+
+/**
+ * @author neo
+ */
+public interface Channel {
+    void send(String message);
+
+    Map<String, Object> context();
+
+    void close();
+}