+++ core-ng/src/main/java/core/framework/web/websocket/ChannelListener.java
@@ -1,0 +1,16 @@
+package core.framework.web.websocket;
+
+import core.framework.web.Request;
+
+/**
+ * @author neo
+ */
+public interface ChannelListener {
+    default void onConnect(Request request, Channel channel) {
+    }
+
+    void onMessage(Channel channel, String message);
+
+    default void onClose(Channel channel) {
+    }
+}