+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### 6.7.1 (8/31/2018 - )
 * http: set tcp keep alive timeout to 620s, to adapt to both AWS ELB and gcloud LB
 * log-processor: fix index template should use strict_date_optional_time as date format
+* ws: draft websocket impl
 
 ### 6.7.0 (8/23/2018 - 8/31/2018)   !!! search API break changes !!!
 * db: skip query.fetch() if limit is 0